Fix for ticket 2581: crash when adding account with typo
If discovery on HTTPS to the given hostname failed, we ended up hitting an error due to attempting to use 'this' in a callback where the value had been overridden; so the app would crash before continuing on to retry on HTTP, then after than to give a failure message. Fixed the references, and it's now correctly giving a friendly app-level error message and giving you a chance to retry.