Internal browser for iphone pretty much working now :D
Includes 'Open in Safari' and 'Post link' options -- the latter can also be triggered with the compose icon on navbar.

Want to add a 'copy link' to copy the URL to clipboard, but this isn't currently available in the titanium mobile api. :P
Would also be nice to have options to open in alternate browsers where those support custom URL schemes. Since I'm not sure we can detect presence, this might require some manual configuration options.