Tickets #2679, #2680: support 'sharing'/forwarding out to other apps.
On Android, using a custom function in our titanium_mobile branch we can use the native facility to select any destination app that supports the ACTION_SEND intent with text/plain. (eg same as hitting 'Share' from the Android Twitter app.)
If our custom function isn't available, we fall back to the default email selector, which won't pick up as many apps.
Once the android_native_refactor branch lands upstream in Titanium Mobile, we should be able to do this direct from the JS code and will be able to do it on stock runtime.

On iPhone, we have a hardcoded list of sinks:
- Mail (via system mail dialog)
- Seesmic (via x-seesmic URL; launches app but then crashes for me)
- Twitter (via tweetie: URL)

Not sure if it's feasible to detect handler support though, eg to just list those apps that are present.

If the apps aren't present, hitting the buttons just does nothing, which is not super. :( It could be nice for power users to be able to select which apps they have or something if it comes to that.