[Desktop] Fix #2715: crash caused by buglet in avatar cache trimming
On Desktop, the list of directory entries comes up as File objects rather than strings; we were incorrectly assuming local filenames and esploding when appending them to the cache directory; this would end up causing errors inside a callback which freaked out the system.
Now checking if we've got a File object already in getAvatarFile and just passing it through if so, so we don't have such issues.